fix: delete ACM certificate resource
Waiting for the ACM certificate to be validated incurs in a Terraform timeout when deploying the infra at AWS from an clean slate. This is because the despite some hosted zone is created at Route 53, the name servers must be manually updated at the domain registrar i.e., IONOS. To solve this issue, we could either further automate the workflow by (1) transfer the root domain registration to AWS Route 53 or (2) use the external registrar's Terraform provider, or else (3) fallback to the manual process of first create the certificate at ACM, then execute the deployment pipeline (i.e., 'Terraform apply'), and finally update the name servers created at Route 53 at the domain registrar. Since we want to keep the root domain at IONOS and we want to avoid having to maintain the external registrar's Terraform provider as well as having to deal with further security handling, we chose option #3.